Fixed invalid URLS
Three separate (simple) issues repeated across all of the elements
produced the hundreds of errors.

1. The ‘back’ button used a relative url that wouldn’t be valid in the
context of the fhir page. Commented out for this release, will be put
back in shr.org updates
2. ‘Skip-navbar-top’ navigation links were unmatched. Matched them.
(Although… I can’t seem to find a way to actually access these links
using either mouse or keyboard, so it may be best to fully remove in
future release)
3. “Frame” url used ‘?’ reference path. This was marked as valid, and
there may a better way of using that reference. However, an identical
effect was met through linking to index.html, as the home page loaded
all of the frames.