Use model classes instead of CIMCORE json to construct components.
When making constraints and elements to put into the modeldoc, use the model classes from shr-models
in their original forms. This results in some changes to attribute paths.
When an element definition contains a constraint on a field's subfield, and that field is declared by this element (not inherited),
show those constraints in the "Properties Constrained" portion of the model doc.