Improve readability of conditionals
Compare with null instead of checking a variable's type is undefined when making cardinality string.
Use simpler conditional when getting code system for "includes code" constraint.
Check value's truthiness instead of using "in" operator when checking if a field has constraints and a constraintsFilter.
Since the full path should not include the names of primitives, directly check if the path piece's namespace is "primitive".