Process and display subset constraints
When an entry's property can be one of several different types, the entry can apply a subset constraint
to limit that property to a subset of those types. The subset must contain at least two types.
Display these constraints with the label "Subtype" and a list of types for the value, including
links to other modeldoc pages when available.